% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.R
\name{adjustBroadPos}
\alias{adjustBroadPos}
\title{Give 'broad' start/end points of gene conversion events}
\usage{
adjustBroadPos(tb, repertoire)
}
\arguments{
\item{tb}{data.frame, output from \code{batchConvertAnalysis}.}

\item{repertoire}{A named vector of strings storing nucleotide sequences observed from a repertoire. The \code{names} attribute of the vector stores the sequence identifiers.}
}
\value{
A data.frame with these two columns added to \code{tb}:
\describe{
  \item{soft_start}{the position given in the \code{start} column, padded with the number of nucleotides given in \code{fiveprime_identical_length}.}
  \item{soft_end}{the position given in the \code{end} column, padded with the number of nucleotides given in \code{threeprime_identical_length}.}
}
}
\description{
This function gives the 'broad' definition of gene conversion event boundaries, i.e. the start/end coordinates given in \code{batchConvertAnalysis} with added nucleotides corresponding to sequence stretches 5'/3' of start/end points identical between the observed repertoire sequence and the pseudogene. The output respects the same numbering scheme, respecting the imposed gaps ('.') in the input \code{repertoire}.
}
