% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterResultsByGeneOrder.R
\name{filterResultsByGeneOrder}
\alias{filterResultsByGeneOrder}
\title{Visualise locations of gene conversion events}
\usage{
filterResultsByGeneOrder(
  results,
  gene_order,
  functional_alelle_col = "allele",
  donor_allele_col = "gene"
)
}
\arguments{
\item{results}{data.frame, output from \code{batchConvertAnalysis}.}

\item{gene_order}{A named vector of numeric/integers with the order of germline genes in the genome. Can be obtained from IMGT using the \code{getGeneOrderTable} function in BrepConvert, and further generating this vector from the gene order table (see package vignette for example). Each element should be named by the corresponding germline gene name.}

\item{functional_alelle_col}{Column name in \code{results} denoting the annotated functional allele for the sequence. (Default: 'allele')}

\item{donor_allele_col}{Column name in \code{results} with semicolon-delimited list of annotated pseudogenes for the detected gene conversion event.}
}
\value{
The dataframe in \code{results} with a new column named 'filtered_pseudogenes' appended. This column contains the pruned list of possible genes as donors for the observed gene conversion event, after accounting for the physical order of immunoglobulin gene on the chromosome.
}
\description{
This function takes output from the \code{batchConvertAnalysis} function, and add a new column "filtered_pseudogenes" which is a pruned version of the \code{donor_allele_col} removing pseudogenes which are physically implausible (because they are 3' of the acceptor gene - they would have been removed from the genome during V(D)J recombination).
}
