% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatBlat.R
\name{formatBlat}
\alias{formatBlat}
\title{Function to format BLAT output}
\usage{
formatBlat(tb, adjustment = 6)
}
\arguments{
\item{tb}{data.frame, output from the \code{blat} function.}

\item{adjustment}{integer, the number of nucleotides which flank the gene conversion events (included for BLAT just to facilitate the alignment to look for seeds). This needs to be taken into account to re-adjust the actual start and end point of conversion events. (default: 6)}
}
\description{
Function to format BLAT output
}
