% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGeneOrderTable.R
\name{formatGeneOrderTable}
\alias{formatGeneOrderTable}
\title{Format Gene Order Table from IMGT into a data frame}
\usage{
formatGeneOrderTable(html_table)
}
\arguments{
\item{html_table}{raw html code output from \code{getGeneOrderTable}.}
}
\value{
A data frame with separate columns indicating immunoglobulin gene names and gene order (as an integer, with 1 being the most 5' gene in the locus).
}
\description{
This function parse the HTML code scraped using the \code{geneGeneOrderTable} function into a data frame.
Due to the fact that IMGT formats these tables in different species slightly differently (they have different number of genome assemblies, haplotypes etc.), the exact number of columns (and their names) depend on the queried species and locus.
}
