% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.R
\name{getAAGeneConversion}
\alias{getAAGeneConversion}
\title{Translate amino acid sequences for gene conversion events}
\usage{
getAAGeneConversion(tb, repertoire, functional)
}
\arguments{
\item{tb}{data.frame, output from \code{batchConvertAnalysis}.}

\item{repertoire}{A named vector of strings storing nucleotide sequences observed from a repertoire. The \code{names} attribute of the vector stores the sequence identifiers.}

\item{functional}{character, filepath to FASTA file containing DNA sequence(s) of the functional V gene allele(s).}
}
\value{
A data.frame with these columns added to \code{tb}:
\describe{
  \item{germline_AA_narrow}{(Amino acid sequence of the region between \code{start} and \code{end}, from the functional germline allele.}
  \item{observed_AA_narrow}{Amino acid sequence of the region between \code{start} and \code{end}, from the observed sequence sampled in the repertoire data.}
  \item{germline_AA_broad}{Amino acid sequence of the region between \code{start} and \code{end} plus the flanking stretches given by \code{fiveprime_identical_length} and \code{threeprime_identical_length}, from the functional germline allele.}
  \item{observed_AA_broad}{(Amino acid sequence of the region between \code{start} and \code{end} plus the flanking stretches given by \code{fiveprime_identical_length} and \code{threeprime_identical_length}, from the observed sequence sampled in the repertoire data.}
}
}
\description{
Translate amino acid sequences for gene conversion events
}
