% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoverage.R
\name{getCoverage}
\alias{getCoverage}
\title{Count gene conversion events which cover/start at every DNA position along the germline}
\usage{
getCoverage(
  db,
  column1,
  column2,
  start_col,
  end_col = NULL,
  type = c("coverage", "start")
)
}
\arguments{
\item{db}{data.frame, output from \code{batchConvertAnalysis}.}

\item{column1}{optional, character corresponding to a column name in \code{db}. If supplied, calculation of coverage will be considered separately for every unique value in \code{column1} of the data frame \code{tb} so that histograms can be plotted separately.}

\item{column2}{optional, character corresponding to a column name in \code{db}. If supplied, calculation of coverage will be considered separately for every unique value in \code{column2} of the data frame \code{tb} so that histograms can be plotted separately.}

\item{start_col}{character, column in \code{db} which gives the start position.}

\item{end_col}{character, column in \code{db} which gives the end position. Not used when \code{type == 'start'}.}

\item{type}{character, either 'coverage' (i.e. how often is a given DNA position included within gene conversion events?) or 'start' (i.e. how often do gene conversion events start at a given DNA position?).}
}
\value{
A data.frame with three columns:
\describe{
  \item{position}{numeric, DNA position}
  \item{value}{numeric, if \code{type =='coverage'} the proportion of gene conversion events which cover the given \code{position}. If \code{type == 'start'}, the proportion of events which begin at the given \code{position}.}
  \item{region}{character, either 'FWR' (framework) or 'CDR'. Follows the standard IMGT numbering; the region is inferred from the \code{position} column.}
}
}
\description{
This function counts how often gene conversion events (partitioned by different combinations of values in \code{column1} and \code{column2} of \code{db}) cover/start at every DNA position along the V gene.
}
