% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEditDistance.R
\name{getEditDistance}
\alias{getEditDistance}
\title{Function to calculate Levenshtein distance between defined sequence stretches of two longer sequences.}
\usage{
getEditDistance(hits, query, subject, method = "levenshtein")
}
\arguments{
\item{hits}{data.frame storing potential hits (donor pseudogenes) that explain a given gene conversion event. Each row corresponds to one definition using a given pseudogene. The boundaries of the pseudogene sequence and the observed repertorie sequence are given.}

\item{query}{character, the full-length V gene sequence of the observed repertoire sequence for which gene conversion events are annotated.}

\item{subject}{\code{DNAStringSet} object containing a set of pseudogene DNA sequences.}

\item{method}{the method used to calculate string distances (default: "levenshtein")}
}
\description{
Function to calculate Levenshtein distance between defined sequence stretches of two longer sequences.
}
