% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneOrderTable.R
\name{getGeneOrderTable}
\alias{getGeneOrderTable}
\title{Fetch order of immunoglobulin genes at the genomic locus}
\usage{
getGeneOrderTable(species, locus)
}
\arguments{
\item{species}{string, common name of the species of interest (see 'Description' for further details).}

\item{locus}{string, immunoglobulin locus (IGH, IGK, IGL) (see 'Description' for further details).}
}
\value{
A data frame with separate columns indicating immunoglobulin gene names and gene order (as an integer, with 1 being the most 5' gene in the locus).
}
\description{
This function fetch from the IMGT website on the order of immunoglobulin genes at a given genomic locus for a user-specified species.
The list of species can be viewed here: https://www.imgt.org/IMGTrepertoire/LocusGenes/ (under "Locus gene order"). The common names of species were expected in this function.
}
