% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdenticalLength.R
\name{getIdenticalLength}
\alias{getIdenticalLength}
\title{Function to calculate the length of identical sequence stretch 5' or 3' of given boundaries of sequence stretches from two longer sequences.}
\usage{
getIdenticalLength(hits, subject, query, type)
}
\arguments{
\item{hits}{data.frame storing potential hits (donor pseudogenes) that explain a given gene conversion event. Each row corresponds to one definition using a given pseudogene. The boundaries of the pseudogene sequence and the observed repertorie sequence are given.}

\item{subject}{\code{DNAStringSet} object containing a set of pseudogene DNA sequences.}

\item{query}{character, the full-length V gene sequence of the observed repertoire sequence for which gene conversion events are annotated.}

\item{type}{character, either '5p' or '3p' denoting the direction for which the identical sequence stretches are considered.}
}
\description{
Function to calculate the length of identical sequence stretch 5' or 3' of given boundaries of sequence stretches from two longer sequences.
}
