% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.R
\name{getNarrowCDNA}
\alias{getNarrowCDNA}
\title{Give 'narrow' cDNA sequence of gene conversion events in the observed and the germline sequences.}
\usage{
getNarrowCDNA(start, end, start_c, end_c, seq_id, repertoire, seq)
}
\arguments{
\item{start}{numeric, start point of gene conversion event. (column \code{start} from output of \code{batchConvertAnalysis})}

\item{end}{numeric, end point of gene conversion event. (column \code{start} from output of \code{batchConvertAnalysis})}

\item{start_c}{numeric, position in the codon (i.e. 1, 2, or 3) for the start point of gene conversion event.}

\item{end_c}{numeric, position in the codon (i.e. 1, 2, or 3) for the end point of gene conversion event.}

\item{seq_id}{character, identifier for the sequence of interest. i.e. the relevant entry in the \code{SeqID} column of the output from \code{batchConvertAnalysis}.}

\item{repertoire}{A named vector of strings storing nucleotide sequences observed from a repertoire. The \code{names} attribute of the vector stores the sequence identifiers.}

\item{seq}{character, an ungapped DNA sequence of the functional allele.}
}
\value{
A vector with two characters. The first item corresponds to the cDNA sequence corresponding to the location of the annotated gene conversion event, taken from \code{functional}. The second item corresponds to the cDNA sequence at the location of the gene conversion event, observed for the given \code{seq_id} in \code{repertoire}.
}
\description{
Give 'narrow' cDNA sequence of gene conversion events in the observed and the germline sequences.
}
