% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopBlat.R
\name{getTopBlat}
\alias{getTopBlat}
\title{Function to rank and obtain donor pseudogenes which best explain a given gene conversion event}
\usage{
getTopBlat(tb, start, end, ntop = 5, adjustment = TRUE)
}
\arguments{
\item{tb}{data.frame storing potential hits (donor pseudogenes) that explain a given gene conversion event generated by running the BLAT program.}

\item{start}{integer, start site of the gene conversion event.}

\item{end}{integer, end site of the gene conversion event.}

\item{ntop}{integer, the number of possibilities to return. (default: 5, i.e. the best 5 candidate donor pseudogenes are returned)}

\item{adjustment}{boolean, does the boundaries on the donor pseudogene sequence need to be adjusted? (default: TRUE)}
}
\description{
Function to rank and obtain donor pseudogenes which best explain a given gene conversion event
}
