% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeHits.R
\name{mergeHits}
\alias{mergeHits}
\title{Function to merge adjacent events explainable by the same donor pseudogene}
\usage{
mergeHits(hits, lut_functional)
}
\arguments{
\item{hits}{data.frame storing annotations of the nominated hits (donor pseudogenes) that explain all gene conversion event generated by running the BLAT program.}

\item{lut_functional}{an \code{IRanges} object storing positions of mismatches between all functional alleles against all pseudogene alleles. This acts as a lookup table to determine whether the functional allele and any given pseudogene share identical nucleotide sequences within a range of interest.}
}
\description{
This function merges adjacent events only if the intervening sequence stretch is COMPLETELY identical between the functional allele and the nominated donor pseudogene. As a result of the merge, the boundaries, edit distance and other annotations of the events are adjusted accordingly.
}
