% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCoverage}
\alias{plotCoverage}
\title{Histograms of gene conversion events}
\usage{
plotCoverage(
  tb,
  repertoire,
  column1 = NULL,
  column2 = NULL,
  type = c("coverage", "start"),
  definition = c("narrow", "broad", "both")
)
}
\arguments{
\item{tb}{data.frame, output from \code{batchConvertAnalysis}.}

\item{repertoire}{A named vector of strings storing nucleotide sequences observed from a repertoire. The \code{names} attribute of the vector stores the sequence identifiers.}

\item{column1}{optional, character corresponding to a column name in \code{tb}. If supplied, calculation of coverage will be considered separately for every unique value in \code{column1} of the data frame \code{tb} so that histograms can be plotted separately.}

\item{column2}{optional, character corresponding to a column name in \code{tb}. If supplied, calculation of coverage will be considered separately for every unique value in \code{column2} of the data frame \code{tb} so that histograms can be plotted separately.}

\item{type}{character, either 'coverage' (i.e. how often is a given DNA position included within gene conversion events?) or 'start' (i.e. how often do gene conversion events start at a given DNA position?).}

\item{definition}{character. Indicates which definition of gene conversion event boundaries are to be used. Can take the values of any one of 'narrow' (the strict start/end points given by \code{batchConvertAnalysis}), 'broad' (the start/end padded with nucleotides 5'/3' to the given start/end) or 'both' (both 'narrow' and 'end' will be considered and two sets of plots will be generated.)}
}
\value{
A ggplot2 object showing the histograms, with x-axis showing the DNA position and y-axis showing the percentage of gene conversion events which cover/start at the given position. Framework (FW) or CDR positions are coloured differently. If \code{column1} and/or \code{column2} are provided separate histograms will be plotted for different data partitions specified along the two given columns present in \code{tb}, arranged side-by-side.
}
\description{
This function visualises histograms showing the coverage or starting point of gene conversion events in a given set of annotation generated from \code{batchConvertAnalysis}, using definition of events.
}
