% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScanGeneConversion.R
\name{ScanGeneConversion}
\alias{ScanGeneConversion}
\title{Wrapper function to scan and annotate gene conversion events}
\usage{
ScanGeneConversion(
  seqname,
  repertoire,
  functional,
  pseudogenes,
  blat_all,
  blat_whole,
  lut,
  gapwidth = 3
)
}
\arguments{
\item{seqname}{character, identifier of a given sequence from \code{repertoire}, for which gene conversion events are to be annotated.}

\item{repertoire}{A named vector of strings storing nucleotide sequences observed from a repertoire. The \code{names} attribute of the vector stores the sequence identifiers.}

\item{functional}{\code{DNAStringSet} object storing nucleotide sequences of the functional alleles of the V gene.}

\item{pseudogenes}{\code{DNAStringSet} object storing nucleotide sequences of the pseudogene alleles of the V gene.}

\item{blat_all}{data.frame storing BLAT results when searching subsequences taken from \code{repertoire} against the pseudogene sequences.}

\item{blat_whole}{data.frame storing BLAT results when searching the full-length repertoire sequences against the pseudogene sequences. This is used to supplement pseudogene matches given in \code{blat_all}}

\item{lut}{A \code{IRanges} object storing positions of mismatches comparing all functional alleles against all pseudogene alleles.}

\item{gapwidth}{integer, the number of nucleotides that separate mismatches to be grouped together as one candidate gene conversion event.}
}
\description{
Wrapper function to scan and annotate gene conversion events
}
